gdk/src/*.hg: Mark classes and functions with GDKMM_API
Mark all the classes and methods/functions (if applicable) with
GDKMM_API so that we can use it to export symbols, if gmmproc is new
enough to apply GDKMM_API to the _WRAP_ENUM, _WRAP_GERROR and _CLASS_*
items.

The check on whether gmmproc is new enough will be in an upcoming
commit.