gtk/src/*.hg: Mark classes and functions with GTKMM_API
Mark all the classes and methods/functions (if applicable) with
GTKMM_API so that we can use it to export symbols, if gmmproc is new
enough to apply GTKMM_API to the _WRAP_ENUM, _WRAP_GERROR, _WRAP_EQUAL*
and _CLASS_* items.

The check on whether gmmproc is new enough will be in an upcoming
commit.