NMake Makefiles: Export symbols without gendef if possible
We check to see whether we have a new-enough gmmproc that is capable of
generating the headers from the .hg templates with function decorations
at places where they need to be.  If it is, we skip building gendef and
use __declspec(dllexport) to export the needed symbols.

Otherwise, we continue to use gendef.exe to export symbols, as we did
before.